Move preflight method to Item class
Item is a good place to host preflight method since it's parent of
both File and Folder and preflight is used only for file update (in File)
and new file upload (in Folder).
